Specializes changelog builder to be markdown only
